Create Device.nut
This code configures the temperature, humidity, light and air pressure sensors
from the EV(Tail) on the electric imp. imp001 was used for this code
back to the LiveControl server. 

Please refer to the electric imp agents codes README file on this folder to better understand 
the script and how to run it.